Dummy version of UltiSnips_FileTypeChanged()
Define a dummy version of the UltiSnips_FileTypeChanged function before
checking for python support.  If that function isn't defined, the
autocommand defined in ftdetect/UltiSnips.vim will cause an error any
time a new file is opened.  That autocommand is setup before the
function is defined, so we can't setup the autocommand only when the
function exists.