Fix tests to work with new version of screen
Debian unstable currently has a pre-release version of screen 4.1.0.
This version changes the "stuff" command so that special characters get
parsed, breaking many of the tests.

Detect if the version of screen being used to run tests does this
parsing, and if so do additional escaping of characters so that tests
again work.

With this change, all tests pass with both old and new versions of
screen.