Leave new scratch buffers in command mode
Modify new_scratch_buffer method to leave the created buffer in command
rather than insert mode.  Currently the only use of this method is for
showing stack traces for errors, when that happens it's much more likely
that the user will want to use command mode to move around than to use
insert mode to add text to the beginning of the buffer.