Provide a buffer local override for the global snippet directories.
This introduces the ability to override the global snippet directories
with a buffer local variable.  When b:UltiSnipsSnippetDirectories is
present in a buffer, it's value overrides the value of
g:UltiSnipsSnippetDirectories.  This allows for easier customization of
the snippet directories for coping with projects that have competing
code style standards, without introducing auto commands into your
configuration.