Add "b" trigger option to JS console snippets
The "b" trigger option will require that the snippet be expanded only if
it is at the "beginning of the line", that is, only whitespace may
precede the tab trigger.

As suggested by @SirVer[1], this commit adds the "b" tab trigger option
to all of the JavaScript console API snippets.

[1]: https://github.com/SirVer/ultisnips/pull/122/files#r8665940