Add snippets for scss files
This commit adds a lot of the basic Sass directives used in Sass script.
Documentation can be found at:

  http://sass-lang.com/documentation/file.SASS_REFERENCE.html

I decided to use regular expressions for the triggers, since typing the
`@` may be too cumbersome to require it to always be included, but is
likely happening due to muscle memory too often to leave it out. Regular
expressions give us the flexibility to have it both ways in this case.