Expand environment variables in UltiSnipsSnippetsDir
I noticed that environment variables don't work in the current version of UltiSnips within the value for `g:UltiSnipsSnippetsDir`.  For example, I tried setting it to

  let g:UltiSnipsSnippetsDir = "$DOTFILES/nvim/UltiSnips"

But it would append the directory after the path to the current file or something weird like that.  This change expands the path at the right time so that it can find the snippet directory correctly.