Add option to disable preview window in a snippet
This is in order to workaround YCM when it has the preview enabled. If
YCM triggers the preview while in a snippet, the current snippet context
is torn down.

In the future it may be useful to gracefully maintain the state in this
case, but for now I believe this is a reasonable workaround.