Fix UltiSnipsSnippetDirectories suggestion in docs (#960)
The variable g:UltiSnipsSnippetDirectories including its documentation was introduced in b47e686. There the documentation read

> let g:UltiSnipsSnippetDirectories=[$HOME.'/.vim/UltiSnips']

This example was later changed in #775, claiming that this variable needs to be a string. This seems to have been done in error, so this change reverts that PR back to the original documentation.

Fixes #815.