Disable auto detecting changes to snippet files on expansion (#967)
Instead, trigger an auto command whenever a .snippets file is saved that will reload the snippets. This is a performance improvement with a loss of functionality: externally (i.e. outside of the current Vim instance) generated snippet files will not be picked up until UltiSnips#RefreshSnippets is called.

I decided to not expose a command for this, since I assume it is rarely useful. I think the command would add confusion of when it was supposed to be ran.

Fixes #932.