Add note about contributing to traceback (#969)
There too many issues with traceback only. While it is good to have traceback, it is sell useful when it is the only info in the issue description. I think that contributing manual should be mentioned here even if we have issue template now. At least further new issues with traceback may follow contributing guide.