Explain why following Unite steps didn't work
If you follow the help setup for Unite, and the binding key (F12) is the
same as UltiSnipsListSnippets, then it doesn't work because we clobber
whatever is bound to UltiSnipsListSnippets in
after/plugin/UltiSnips_after.vim. We can skip this clobbering by
setting this flag.

I don't understand why UltiSnips_after is required to make supertab
work, but it was introduced to fix #212. As a better fix to #212, I'd
suggest checking for supertab or if supertab is bound to the expected
key sequence instead of always stomping the bindings.