If UltiSnipsSnippetDirectories only contains a single directory, expand `~` (#1085)
Fixes #948. The issue was that there was an early termination condition that, if hit, meant the specified directories would never have ~ expansion performed, so absolute paths had to be specified in the config variables.

With this patch, ~ can be used in the config and will be parsed properly under both conditions.