Use blackhole register for select-mode deletion (#1050)
This simple PR is for having SELECT-mode mappings not interfere with the clipboard register.
Since Ultisnips basically assumes control over SELECT mode w.r.t. some mappings, imposing a life cycle. There is nothing the user can do about it. When executing a snippet with a default placeholder and whishing to <Del>ete it in place, the local clipboard will inevitably be overwritten with the default placeholder.