Do not use realpath in most cases (#1154)
If a user has created symlinks, we have no real reason to abstract back to the real inode path.

Side effect of this could be that the same snippet file shows up multiple times in `:UltiSnipsEdit`, but under different paths if symlinks are involved. This however feels like it is working as intended.

Fixes #1153.