Send the correct trigger key to _handle_failure()
Currently, "SnippetManager._handle_failure()" receives the backward
trigger key instead of the list trigger key when listing fails
because there are no snippets for the current buffer.

If "g:UltiSnipsJumpBackwardTrigger" is set to, for example
"<S-Tab>", failing to list any snippets will send that key to the
type-ahead instead of the value of "g:UltiSnipsListSnippets".

In my Vim configuration, I have the following settings:

  let g:UltiSnipsJumpBackwardTrigger = '<S-Tab>'
  let g:UltiSnipsListSnippets = '<C-L>'

I get the following behavior before this commit:

  1. Enter Insert mode
  2. Press <C-L> to list snippets, listing fails b/c no snippets
  3. A tab character is inserted when nothing should happen

This commit uses _handle_verify() properly so that no useless
characters are inserted into the buffer when there are no snippets.