feat(matchit): add pattern for interpolation
This commit makes editing interpolation in snippets easier with the
plugin, vim-matchup (https://github.com/andymass/vim-matchup).

```snippets
`!p fo|o` # '|' is cursor position
```

  |
 ci%
  ↓

```snippets
`!p |`
```

Note:
It's undesirable to join 'snip.rv' in b:match_words as a middle pattern
because it often disturbs to `ci%` to edit an interpolation, or even to
`da%` to delete/move a whole interpolation (usually, b:match_words
doesn't include '`:`' because a backtick would be used not as a pair).