Added CanExpandSnippet, CanJumpForwards, CanJumpBackwards (#1314)
Closes #786.

Adds three functions, `UltiSnips#CanExpandSnippet`, `UltiSnips#CanJumpForwards`, and `UltiSnips#CanJumpBackwards`, which cleans up more complex expansion functions.

The issue with the current method is that it requires we actually *run* the `<C-R>=UltiSnips#ExpandSnippet()<CR>` then check the result. This makes actually doing anything after it very difficult if you're in an `<expr>` mapping. Now we can have `<expr>` mappings that look like this: 

```lua
function expand_or_jump()
    -- ...
    if vim.api.nvim_eval([[ UltiSnips#CanExpandSnippet() ]]) == 1 then
        return [[ <C-R>=UltiSnips#ExpandSnippet()<CR> ]]
    end
    -- ...
end
```

and etc.

Issue #786 was previously closed because it was said that `UltiSnips#SnippetsInCurrentScope()` has the same functionality, but it actually doesn't as I pointed out in the comments.