Use a valid unite kind variable (#1338)
Fix #1337.

Unite defines a list of valid values for 'kind' in `:help unite-kind-variables`. UltiSnips was using a custom one, but
doesn't seem to be adding a custom implementation for it. Instead, use the default 'common' one. This allows
s:unite_source.action_table.expand.func to be called when the snippet is invoked.

Test
Follow steps and minimal vimrc on issue #1337: invoke unite and expand first snippet.

Partially reverts #1043, but that seems correct as confirmed by original author @weilbith.