Fix #1339 wrong cursor position in nested snippets (#1347)
This PR fixes the problem mentioned in issue #1339. _VimCursor dummy NoneditableTextObject is now created as a child of the EditableTextObject actually being edited at that time. Thereby it is ensured that it is unambiguous which mirrors should and should not move the cursor position further. It also resolves general issues when a snippet contains only tabstops and mirrors and no actual text.

Tests have been added to ensure that this remains functional and the test suite completes successfully.

Fixes #1339.