Check success of "tmux send-keys"
If for any reason tmux failed, reveal the error raising a subprocess exception.
This makes debugging a lot easier. If something goes wrong, for instance tmux
session is not started, we will see stack trace, command arguments as an
array, and all output. However, the output is printed in a mess with unittest
messages, but it is okay.

Furthermore, the Python documentation cautions against using stdout=PIPE or
stderr=PIPE with the call() function.