fix #1165: insert mode lag
Ultisnips was traversing the whole runtimepath every time an expansion
was attempted, something that happens for every key press since there is
autotrigger.

Even without autotrigger, traversal was done at every snippet expansion,
which caused slowdowns in that case.

With this change, traversal happens only the first time, or after
a source must be refreshed (because a snippets file has been updated).